Add langchain based RAG pipeline
